Modularize Validator component to component and plugin
